move: auth dir to src/
