--- user-service/src/api/v1/auth/login/dto/login-requestdto.ts
@@ -1,6 +1,0 @@
-import { PickType } from '@nestjs/swagger';
-import { User } from '../../../user/entity/user.entity';
-
-export class LoginRequestDto extends PickType(User, [
-    'phoneNumber', 'nickName'
-]) {}
