+++ user-service/src/app.module.ts
@@ -1,15 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import databaseConfiguration from 'config/database.configuration';
-import { AuthModule } from './api/v1/auth/auth.module';
-import { AuthService } from './api/v1/auth/auth.service';
+import { AuthModule } from './auth/auth.module';
+import { AuthService } from './auth/auth.service';
 import { LikeModule } from './api/v1/like/like.module';
 import { SearchModule } from './api/v1/search/search.module';
-import { UserModule } from './api/v1/user/user.module';
 import { VideoModule } from './api/v1/video/video.module';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { DatabaseModule } from './database/database.module';
+import { UserModule } from './user/user.module';
 
 @Module({
   imports: [
@@ -24,7 +22,5 @@ import { DatabaseModule } from './database/database.module';
     VideoModule,
     DatabaseModule
   ],
-  controllers: [AppController],
-  providers: [AppService, AuthService],
 })
 export class AppModule {}
