+++ user-service/src/auth/dto/login-requestdto.ts
@@ -1,0 +1,8 @@
+export class LoginRequestDto {
+    private phoneNumber: string;
+    private nickName: string;
+
+    getPhoneNumber(){ return this.phoneNumber; }
+
+    getNickName() { return this.nickName; }
+}
