+++ user-service/src/auth/dto/login-response.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { CommonResponseDto } from "../../../swagger/dto/common-response.dto";
-import { UserDto } from "../../../user/dto/user.dto";
+import { CommonResponseDto } from "../../api/v1/swagger/dto/common-response.dto";
+import { UserDto } from "../../user/dto/user.dto";
 
 class LoginResponseData {
     @ApiProperty()
